+++ cli/csstools-cli/src/cli.ts
@@ -10,6 +10,7 @@ import postcssDoublePositionGradients from './plugins/postcss-double-position-gr
 import postcssEnvFunction from './plugins/postcss-env-function';
 import postcssFocusVisible from './plugins/postcss-focus-visible';
 import postcssFocusWithin from './plugins/postcss-focus-within';
+import postcssFontFormatKeywords from './plugins/postcss-font-format-keywords';
 import postcssGapProperties from './plugins/postcss-gap-properties';
 import postcssHWBFunction from './plugins/postcss-hwb-function';
 import postcssImageSetFunction from './plugins/postcss-image-set-function';
@@ -17,6 +18,7 @@ import postcssIsPseudoClass from './plugins/postcss-is-pseudo-class';
 import postcssLabFunction from './plugins/postcss-lab-function';
 import postcssLogical from './plugins/postcss-logical';
 import postcssNesting from './plugins/postcss-nesting';
+import postcssNormalizeDisplayValues from './plugins/postcss-normalize-display-values';
 import postcssOverflowShorthand from './plugins/postcss-overflow-shorthand';
 import postcssPlace from './plugins/postcss-place';
 import postcssPresetEnv from './plugins/postcss-preset-env';
@@ -61,6 +63,9 @@ function main() {
 		case 'postcss-focus-within':
 			postcssFocusWithin();
 			return;
+		case 'postcss-font-format-keywords':
+			postcssFontFormatKeywords();
+			return;
 		case 'postcss-gap-properties':
 			postcssGapProperties();
 			return;
@@ -81,6 +86,9 @@ function main() {
 			return;
 		case 'postcss-nesting':
 			postcssNesting();
+			return;
+		case 'postcss-normalize-display-values':
+			postcssNormalizeDisplayValues();
 			return;
 		case 'postcss-overflow-shorthand':
 			postcssOverflowShorthand();
@@ -118,6 +126,7 @@ function main() {
 				'  postcss-env-function',
 				'  postcss-focus-visible',
 				'  postcss-focus-within',
+				'  postcss-font-format-keywords',
 				'  postcss-gap-properties',
 				'  postcss-hwb-function',
 				'  postcss-image-set-function',
@@ -125,6 +134,7 @@ function main() {
 				'  postcss-lab-function',
 				'  postcss-logical',
 				'  postcss-nesting',
+				'  postcss-normalize-display-values',
 				'  postcss-overflow-shorthand',
 				'  postcss-place',
 				'  postcss-pseudo-class-any-link',