+++ rollup/configs/externals.js
@@ -3,8 +3,10 @@ export const externalsForCLI = [
 	'path',
 	'url',
 
+	'@csstools/postcss-font-format-keywords',
 	'@csstools/postcss-hwb-function',
 	'@csstools/postcss-is-pseudo-class',
+	'@csstools/postcss-normalize-display-values',
 	'autoprefixer',
 	'browserslist',
 	'caniuse-lite',